# lib/api-zod/src/generated/types/chordEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MusicAI Studio - AI-powered music intelligence and generation API
 * OpenAPI spec version: 0.1.0
 */

export interface ChordEvent {
  startTime: number;
  endTime: number;
  /** e.g. Cmaj7, Am, G7/B */
  chord: string;
  romanNumeral: string;
  confidence: number;
  alternatives?: string[];
}