# lib/api-zod/src/generated/types/exportRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MusicAI Studio - AI-powered music intelligence and generation API
 * OpenAPI spec version: 0.1.0
 */
import type { ExportRequestFormatsItem } from "./exportRequestFormatsItem";

export interface ExportRequest {
  /** @minItems 1 */
  formats: ExportRequestFormatsItem[];
}