# lib/api-zod/src/generated/types/keyAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MusicAI Studio - AI-powered music intelligence and generation API
 * OpenAPI spec version: 0.1.0
 */
import type { KeyAnalysisMode } from "./keyAnalysisMode";
import type { Modulation } from "./modulation";

export interface KeyAnalysis {
  /** e.g. C, Am, F#m */
  globalKey: string;
  mode: KeyAnalysisMode;
  confidence: number;
  modulations?: Modulation[];
}