# lib/api-zod/src/generated/types/melodyAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MusicAI Studio - AI-powered music intelligence and generation API
 * OpenAPI spec version: 0.1.0
 */
import type { MelodyNote } from "./melodyNote";

export interface MelodyAnalysis {
  notes: MelodyNote[];
  /** Top 3 inferred chord progressions from melody */
  inferredHarmony?: string[];
}